requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
email-validator==2.1.0.post1
pyswisseph==2.10.3.2
weasyprint
Jinja2==3.1.2
geopy==2.4.0